+++ catalogue/Catalogue.hpp
@@ -659,6 +659,18 @@ public:
   virtual common::dataStructures::VidToTapeMap getTapesByVid(const std::set<std::string> &vids) const = 0;
 
   /**
+   * Returns the tapes with the specified volume identifiers.
+   *
+   * This method will throw an exception if it cannot find ALL of the specified
+   * tapes.
+   *
+   * @param vids The tape volume identifiers (VIDs).
+   * @param ignoreMissingVids Allow non-existing VIDs to be ignored.
+   * @return Map from tape volume identifier to tape.
+   */
+  virtual common::dataStructures::VidToTapeMap getTapesByVid(const std::set<std::string> &vids, bool ignoreMissingVids) const = 0;
+
+  /**
    * Returns map from VID to logical library name for specified set of VIDs.
    *
    * @param vids The tape volume identifiers (VIDs).